Add .vimrc with settings compatible with VS.
VS reformats the XML of any property sheets you use. So if you edit it with a
text editor, you better match exactly the format how VS writes the property
sheets, or it will modify them and git diff will show it as a change to commit.
This sets the appropriate settings in Vim for tab expansion etc. to be
compatible with Visual Studio.