Added basename naming - specifying nonmangled part.
naming\basename enables you to specify nonmangled part of your binary. This is
orthogonal to mangling.

There are 2 "sections" of the nonmangled name - the solution section
(SolutionPart) and project section (ProjectPart).

The "format" policy defines whether both of these parts are used and how they
are concatenated. The names of the property sheets should be self-explanatory.

The "case" policy defines whether the parts will be CamelCased, lowercased, or
in their original case.

For camel and lower case, you have to do a bit hand-work by defining macros
$(ProjectNameCamel) etc., and the same for solution part. I hope MSBuild has
such functions, or is at least extensible in .NET in order to be able to define
lowercasing and CamelCasing functions so that the user does not have to do it
by hand in future MSBuild-based VCConf.